--Can now handle large int with (long long)

+++ TheMaximumPairwiseProductProgrammingChallenge/maximimPairwiseProductChallengeIn.cpp
@@ -6,10 +6,10 @@ using std::vector;
 using std::cin;
 using std::cout;
 
-int MaxPairwiseProduct(const vector<int>& numbers) {
+long long MaxPairwiseProduct(const vector<int>& numbers) {
 
     // Initialize result
-    int result = 0;
+    long long result = 0;
     int n = numbers.size();
 
     // Nested Loop
@@ -17,8 +17,8 @@ int MaxPairwiseProduct(const vector<int>& numbers) {
             // Then compare to see if the element is the biggest thus far
     for (int i = 0; i < n; ++i) {
         for (int j = i + 1; j < n; ++j) {
-            if (numbers[i] * numbers[j] > result) {
-                result = numbers[i] * numbers[j];
+            if (((long long)numbers[i]) * numbers[j] > result) {
+                result = ((long long)numbers[i]) * numbers[j];
             }
         }
     }
@@ -34,7 +34,7 @@ int main() {
         cin >> numbers[i];
     }
 
-    int result = MaxPairwiseProduct(numbers);
+    long long result = MaxPairwiseProduct(numbers);
     cout << result << "\n";
     return 0;
 }
